bigquery-heatwave-py/requirements.txt
fdk>=0.1.91
oci
mysql-connector-python